+++ src/main/java/com/glisco/numismaticoverhaul/currency/CurrencyResolver.java
@@ -8,8 +8,8 @@ public class CurrencyResolver {
      * @param rawValue The value to split
      * @return The individual currency values in the format int[]{BRONZE, SILVER, GOLD}
      */
-    public static int[] splitValues(int rawValue) {
-        int[] output = new int[]{0, 0, 0};
+    public static long[] splitValues(long rawValue) {
+        long[] output = new long[]{0, 0, 0};
 
         if (rawValue / 10000 != 0) {
             output[2] = rawValue / 10000;
@@ -34,7 +34,7 @@ public class CurrencyResolver {
      * @param values The individual currency values in the format int[]{BRONZE, SILVER, GOLD}
      * @return The raw value, with respect to each value's worth
      */
-    public static int combineValues(int[] values) {
+    public static long combineValues(long[] values) {
         if (values.length != 3) throw new IllegalArgumentException("Input array has to have 3 elements");
 
         return values[0] + values[1] * 100 + values[2] * 10000;
