+++ src/main/java/com/glisco/numismaticoverhaul/item/CurrencyItem.java
@@ -4,12 +4,12 @@ import net.minecraft.item.ItemStack;
 
 public interface CurrencyItem {
 
-    static void setOriginalValue(ItemStack stack, int value) {
-        stack.getOrCreateNbt().putInt("OriginalValue", value);
+    static void setOriginalValue(ItemStack stack, long value) {
+        stack.getOrCreateNbt().putLong("OriginalValue", value);
     }
 
-    static int getOriginalValue(ItemStack stack) {
-        return stack.getOrCreateNbt().getInt("OriginalValue");
+    static long getOriginalValue(ItemStack stack) {
+        return stack.getOrCreateNbt().getLong("OriginalValue");
     }
 
     static boolean hasOriginalValue(ItemStack stack) {
@@ -18,8 +18,8 @@ public interface CurrencyItem {
 
     boolean wasAdjusted(ItemStack other);
 
-    int getValue(ItemStack stack);
+    long getValue(ItemStack stack);
 
-    int[] getCombinedValue(ItemStack stack);
+    long[] getCombinedValue(ItemStack stack);
 
 }
