+++ src/main/java/com/glisco/numismaticoverhaul/mixin/TradeOfferMixin.java
@@ -51,11 +51,11 @@ public class TradeOfferMixin implements NumismaticTradeOfferExtensions {
 
         if (!(this.firstBuyItem.getItem() instanceof CurrencyItem currencyItem)) return;
 
-        int originalValue = currencyItem.getValue(this.firstBuyItem);
-        int adjustedValue = numismatic$reputation < 0 ?
-                (int) (originalValue + Math.abs(numismatic$reputation) * (Math.abs(originalValue) * .02))
+        long originalValue = currencyItem.getValue(this.firstBuyItem);
+        long adjustedValue = numismatic$reputation < 0 ?
+                (long) (originalValue + Math.abs(numismatic$reputation) * (Math.abs(originalValue) * .02))
                 :
-                (int) Math.max(1, originalValue - Math.abs(originalValue) * (numismatic$reputation / (numismatic$reputation + 100f)));
+                (long) Math.max(1, originalValue - Math.abs(originalValue) * (numismatic$reputation / (numismatic$reputation + 100f)));
 
         adjustedValue = Math.min(adjustedValue, 990000);
